+++ packages/terra-data-grid/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Fixed
   * Fixed issue where focus was given to the column header instead of its button element.
+ 
+* Added
+  * Added `hasVisibleColumnHeaders` prop for FlowsheetDataGrid to toggle visibility of column headers.
 
 ## 1.1.0 - (November 9, 2023)