+++ packages/terra-framework-docs/CHANGELOG.md
@@ -4,8 +4,12 @@
 
 * Added
   * Added new test for Terra Slider component for long field labels.
+  * Added `hasVisibleColumnHeaders` example for FlowsheetDataGrid.
+  
 * Updated
   * Updated About page of `terra-table` to provide accessibility documentation for Home and End keys.
+  * Removed alpha notice for FlowsheetDataGrid.
+
 
 ## 1.44.0 - (November 9, 2023)